use transform cache for pose transforms as well
